fix(gateway): enforce rolling usage caps and bill abandoned streams
Two money holes on the endpoint desktop, CLI and the VS Code extension
use. Both were verified against source before changing anything.

ROLLING CAPS — the gateway called the legacy 8-argument
`reserve_managed_usage_request(...)` with no ceilings, while apps/web
called `reserve_managed_usage_request_with_limits(...)` with all three.
The capped function is a wrapper that delegates to the legacy one after
checking, so the gateway was literally "the same reservation, minus every
cap". A single user could burn an entire monthly allocation inside one
five-hour window, all of it on flagship models, and nothing on this path
would decline it.

The cause was structural, not an oversight: the ceiling policy lived in
apps/web/lib/server/managed-usage-policy.ts behind `import 'server-only'`,
so the gateway could not reach it. Duplicating the table would have
recreated the drift, so the policy moved to @agiworkforce/types next to
billing-catalog.ts — one definition both surfaces read. apps/web
re-exports every symbol, so nothing there changed.

`is_flagship` matters beyond the request that carries it: it is stamped
onto settlement metadata only inside `_with_limits`, and the flagship
weekly window filters on that tag. Reserving through the legacy function
left gateway spend invisible to the flagship ceiling even for later
apps/web checks.

ABANDONED STREAMS — a client that walked away after receiving tokens
settled as `failed`, which forces actual_cost_cents to 0 and refunds the
whole reservation. "Read the answer, then close the socket" therefore
cost nothing while the provider had already charged us. It was unbounded
rather than merely cheap: a zero settle records no deduction, so it never
counted toward the rolling windows either. Only an attempt that produced
NOTHING is still released in full.

TESTS — the existing reservation test asserted the legacy function name,
so it stayed green for exactly as long as the gateway enforced no caps;
it now pins `_with_limits` and asserts the ceilings are non-zero. The
existing disconnect test passes only because its adapter emits no chunks,
so a `deltas-then-client-abort` mode and a test were added. That test was
confirmed to FAIL against the old settlement line before the fix was
restored — a regression test that passes either way is worthless.

Gateway 276 pass; web 6,549 pass; both typecheck clean.
